Handle overflow when writing events to the file in V2 trace format
We assumed the buffer was large enough to hold all encoded events when
processing trace events. For V1 this was possible because the event
entry is fixed size and it is possible to allocate large enough
buffers. For format V2, we use variable length encoding and hence the
actual size required would be much smaller than the worst case estimate.
It is not good to further increase the size of buffers, so update the
implementation to handle overflows. It just finished the events block
when the buffer is almost full and starts a new one. We don't expect
this to happen often, so this shouldn't have any noticeable performance
/ file size impact.

Bug: 259258187
Test: art/test.py -t 2246-trace-v2
Change-Id: Id6b128dd3b9de33301c7313f5baf8cf9631077ef