Count SDM files into TYPE_DEXOPT_ARTIFACT in ART-managed file stats.
SDM files contain ODEX and ART files, which are dexopt artifacts.

We don't need to count the DM files containing VDEX files because the
storage manager counts them itself.

This CL implements the extension to getUsableArtifacts to achieve this,
which is also necessary for the file GC in the next CL.

Also:
- Fix a bug that the runtime images were not counted if the artifact
  location is ArtifactsLocation.DM.
- Replace @AutoValue classes with records, to simplify the code.

Bug: 377474232
Test: atest ArtServiceTests
Test: atest art_standalone_artd_tests
Change-Id: If1d05d69051d09ebd7f10e6f487cd7892590921c